Create README.md with Installation instructions
For brand new would-be julia users who want to get started with Julia, I would love to send them to this github repo to get started, but there's no information on this page explaining _what to do once you're here!_

I know that these tutorials were originally written with JuliaBox in mind, so i've tried to maintain that spirit in these instructions. But hopefully we can also include instructions here about how to get started with julia in _any capacity_, because JuliaBox may not be the preferred option for everyone, and these are great tutorials that it would be a shame to not share! :)