Add jupyter configuration files to export .jl files
Add Jupyter configs to automatically export a script file whenever a
Jupyter Notebook file (.ipynb) in the repo is saved. This makes diffing
changes in git significantly easier.

The files are exported to a top-level directory called .nbexports, which
mirrors the filepaths of the real files. (Note, when scripts are renamed
/ deleted, it will require manual cleanup from the .nbexports directory)