Clarify the Euler's Number constant in the sigmoid function example
It is better to be explicit with the Euler's Number in the equation
and add a short explanation and a way to use the constant in Julia
than to use `expr` which is inconsistent with the code that followed.

Also correct 'e' to unicode `e` in the code followed (consider using `expr` here
if it makes the code clearer).

Minor: format `g` function name as code.